python-rag-server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sentence-transformers==2.2.2
qdrant-client==1.7.0
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
markdown==3.5.1
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
